requirements.txt
fastapi
uvicorn
pypdf
chromadb
python-multipart
requests
pillow
pytesseract
